cur_sample is float in grain, prevent interpolation from going over the total number of samples in the buffer in grain

+++ Grain.h
@@ -39,7 +39,7 @@ class Grain {
     //Voice Housekeeping
     bool note_on{false};
     float elapsed_ms{0.0f};
-    int cur_sample{0};
+    float cur_sample{0.0f};
 
     //Grain Parameters
     float grain_size; // length of grain in seconds
